errors with bmi calc

+++ src/main/java/com/example/itrack/math/CalCalc.java
@@ -1,18 +1,19 @@
 package com.example.itrack.math;
 
 public class CalCalc {
-    private int weight;
-    private int height;
+    private double weight;
+    private double height;
+
     private int age;
     private double tdee;
     private int totalCal;
 
-    public CalCalc(int age, String gender, int height, int weight, int goalWeight, String activityLevel) {
+    public CalCalc(int age, String gender, double height, double weight, double goalWeight, String activityLevel) {
         this.age = age;
         this.height = height;
         this.weight = weight;
         this.tdee = calculateTDEE(gender, activityLevel);
-        this.totalCal = calculateTotalCal(goalWeight);
+        this.totalCal = calculateTotalCal((int) goalWeight);
     }
 
     private double calculateBMR(String gender) {
@@ -54,19 +55,19 @@ public class CalCalc {
 
     // Getters and other methods...
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         this.height = height;
     }
